client/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
fastjsonschema==2.20.0
Flask==3.0.3
frozenlist==1.4.1
gotrue==2.8.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
multidict==6.1.0
nbformat==5.10.4
packaging==24.1
platformdirs==4.3.6
pluggy==1.5.0
postgrest==0.16.11
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
realtime==2.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
websockets==12.0
Werkzeug==3.0.4
yarl==1.11.1